DiscordBot/Discord.py/requirements.txt
discord.py
python-decouple
sentry-sdk